# PERF: Improve performance of where when using an array of values