--- a/activerecord/lib/arel/table.rb
+++ b/activerecord/lib/arel/table.rb
@@ -98,6 +98,7 @@ module Arel # :nodoc: all
 
     def type_cast_for_database(attribute_name, value)
       type_caster.type_cast_for_database(attribute_name, value)
+    rescue ::RangeError
     end
 
     def able_to_type_cast?