--- a/activerecord/lib/arel/visitors/to_sql.rb
+++ b/activerecord/lib/arel/visitors/to_sql.rb
@@ -81,6 +81,10 @@ module Arel # :nodoc: all
           end
         end
 
+        def visit_Arel_Nodes_CastedArray(o, collector)
+          collector << o.value_for_database.map! { |v| quote(v) }.join(", ")
+        end
+
         def visit_Arel_Nodes_Casted(o, collector)
           collector << quote(o.value_for_database).to_s
         end
@@ -512,12 +516,8 @@ module Arel # :nodoc: all
         def visit_Arel_Nodes_In(o, collector)
           attr, values = o.left, o.right
 
-          if Array === values
-            unless values.empty?
-              values.delete_if { |value| unboundable?(value) }
-            end
-
-            return collector << "1=0" if values.empty?
+          if Arel::Nodes::CastedArray === values
+            return collector << "1=0" if values.value.empty?
           end
 
           visit(attr, collector) << " IN ("
@@ -527,12 +527,8 @@ module Arel # :nodoc: all
         def visit_Arel_Nodes_NotIn(o, collector)
           attr, values = o.left, o.right
 
-          if Array === values
-            unless values.empty?
-              values.delete_if { |value| unboundable?(value) }
-            end
-
-            return collector << "1=1" if values.empty?
+          if Arel::Nodes::CastedArray === values
+            return collector << "1=1" if values.value.empty?
           end
 
           visit(attr, collector) << " NOT IN ("